testing/web-platform/tests/2dcontext/drawing-images-to-the-canvas/drawimage_html_image_4_srgb.html

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<canvas id="dest" height="100" width="100"></canvas>

<script>
var sourceWidth = 100;
var sourceHeight = 100;
var smoothingEnabled = false;
var destCanvas = document.getElementById('dest');
var sourceImg = document.createElement('img');
sourceImg.src = '../2x2-srgb.png'
sourceImg.width = sourceWidth;
sourceImg.height = sourceHeight;
var destCtx = destCanvas.getContext('2d');

function checkPixel(x, y, expected, epsilon) {
    var actual = destCtx.getImageData(x, y, 1, 1).data;
    assert_true(actual.length === expected.length);
    for (var i=0; i < actual.length; i++)
        assert_approx_equals(actual[i], expected[i], epsilon);
}

var pixelTests = [
    ['Pixel Top Left Canvas Background', 0, 0, [255, 0, 0, 255], 1],
    ['Pixel Top Right Canvas Background', 0, 99, [255, 0, 0, 255], 1],
    ['Pixel Bottom Left Canvas Background', 99, 0, [255, 0, 0, 255], 1],
    ['Pixel Bottom Right Canvas Background', 99, 99, [255, 0, 0, 255], 1],

    ['Pixel Surrounding Top Left Canvas Background', 49, 49, [255, 0, 0, 255], 1],
    ['Pixel Surrounding Top Right Canvas Background', 49, 66, [255, 0, 0, 255], 1],
    ['Pixel Surrounding Bottom Left Canvas Background', 66, 49, [255, 0, 0, 255], 1],
    ['Pixel Surrounding Bottom Right Canvas Background', 66, 66, [255, 0, 0, 255], 1],

    ['Pixel Top Left Image', 50, 50, [253, 140, 245, 255], 1],
    ['Pixel Top Right Image', 50, 65, [41, 122, 115, 255], 1],
    ['Pixel Bottom Left Image', 65, 50, [41, 122, 115, 255], 1],
    ['Pixel Bottom Right Image', 65, 65, [253, 140, 245, 255], 1],
];

function runTest() {
    destCtx.fillStyle = "#FF0000";
    destCtx.fillRect(0, 0,  destCanvas.width, destCanvas.height);
    destCtx.imageSmoothingEnabled = smoothingEnabled;
    // 4 arguments, the dest origin is not 0,0 and the dest size is provided but
    // does not match the size of the source. The image will be distorted
    // The source canvas will copied to the 50,50 position of the destination canvas
    // and it will be shrunk to a and area of 16x16
    destCtx.drawImage(sourceImg, 50, 50, 16, 16);
    generate_tests(checkPixel, pixelTests);
}

async_test(t => {
        window.onload = function() {
            t.step(runTest);
            t.done();
        }
}, 'Draw 100x100 image to 100x100 canvas at 50,50 with resize to 16x16.');
</script>
